(WIP) Starting on the API refactor:
* moving things out of miscellaneous files (constants.py) into themed modules
* moving non-public packages (e.g. objects) closer to things that depend on them (e.g checkout/commit code)
* extracting hooks (sgfile commands, custom objects formats, custom mount handlers) up
