+++ splitgraph/commands/commit.py
@@ -3,6 +3,7 @@ from random import getrandbits
 
 from psycopg2.sql import SQL, Identifier
 
+from splitgraph.commands.objects.creation import record_table_as_diff, record_table_as_snap
 from splitgraph.connection import get_connection
 from splitgraph.meta_handler.common import ensure_metadata_schema
 from splitgraph.meta_handler.images import add_new_image
@@ -10,14 +11,13 @@ from splitgraph.meta_handler.misc import table_schema_changed
 from splitgraph.meta_handler.objects import register_table
 from splitgraph.meta_handler.tables import get_table
 from splitgraph.meta_handler.tags import get_current_head, set_head
-from splitgraph.objects.creation import record_table_as_diff, record_table_as_snap
-from splitgraph.pg_audit import manage_audit_triggers, discard_pending_changes
 from splitgraph.pg_utils import get_all_tables
+from ._pg_audit import manage_audit_triggers, discard_pending_changes
 
 
 def commit(repository, image_hash=None, include_snap=False, comment=None):
     """
-    Commits all pending changes to a given mountpoint, creating a new image.
+    Commits all pending changes to a given repository, creating a new image.
 
     :param repository: Repository to commit.
     :param image_hash: Hash of the commit. Chosen by random if unspecified.
