+++ splitgraph/connection.py
@@ -1,3 +1,4 @@
+import re
 from contextlib import contextmanager
 
 import psycopg2
@@ -29,3 +30,19 @@ def get_connection():
     if not _PSYCOPG_CONN:
         _PSYCOPG_CONN = psycopg2.connect(POSTGRES_CONNECTION)
     return _PSYCOPG_CONN
+
+
+def parse_connection_string(conn_string):
+    """
+    :return: a tuple (server, port, username, password, dbname)
+    """
+    match = re.match(r'(\S+):(\S+)@(.+):(\d+)/(\S+)', conn_string)
+    return match.group(3), int(match.group(4)), match.group(1), match.group(2), match.group(5)
+
+
+def serialize_connection_string(server, port, username, password, dbname):
+    return '%s:%s@%s:%s/%s' % (username, password, server, port, dbname)
+
+
+def make_conn(server, port, username, password, dbname):
+    return psycopg2.connect(host=server, port=port, user=username, password=password, dbname=dbname)
