+++ splitgraph/constants.py
@@ -1,7 +1,5 @@
 import logging
-import re
 from collections import namedtuple
-from random import getrandbits
 
 from splitgraph.config import CONFIG
 
@@ -22,42 +20,6 @@ S3_ACCESS_KEY = CONFIG["SG_S3_KEY"]
 S3_SECRET_KEY = CONFIG["SG_S3_PWD"]
 
 
-class SplitGraphException(Exception):
-    pass
-
-
-def get_random_object_id():
-    """Assign each table a random ID that it will be stored as. Note that postgres limits table names to 63 characters,
-    so the IDs shall be 248-bit strings, hex-encoded, + a letter prefix since Postgres doesn't seem to support table
-    names starting with a digit."""
-    return "o%0.2x" % getrandbits(248)
-
-
-def parse_connection_string(conn_string):
-    """
-    :return: a tuple (server, port, username, password, dbname)
-    """
-    match = re.match(r'(\S+):(\S+)@(.+):(\d+)/(\S+)', conn_string)
-    return match.group(3), int(match.group(4)), match.group(1), match.group(2), match.group(5)
-
-
-def serialize_connection_string(server, port, username, password, dbname):
-    return '%s:%s@%s:%s/%s' % (username, password, server, port, dbname)
-
-
-class Color:
-    PURPLE = '\033[95m'
-    CYAN = '\033[96m'
-    DARKCYAN = '\033[36m'
-    BLUE = '\033[94m'
-    GREEN = '\033[92m'
-    YELLOW = '\033[93m'
-    RED = '\033[91m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-    END = '\033[0m'
-
-
 class Repository(namedtuple('Repository', ['namespace', 'repository'])):
     """
     A repository object that encapsulates the namespace and the actual repository name.
@@ -75,5 +37,4 @@ def to_repository(schema):
     if '/' in schema:
         namespace, repository = schema.split('/')
         return Repository(namespace, repository)
-    else:
-        return Repository('', schema)
+    return Repository('', schema)
