+++ splitgraph/drawing.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 from collections import defaultdict
 
-from splitgraph.constants import SplitGraphException
+from splitgraph.exceptions import SplitGraphException
 from splitgraph.meta_handler.common import ensure_metadata_schema
 from splitgraph.meta_handler.images import get_all_images_parents
 from splitgraph.meta_handler.tags import get_current_head
