+++ splitgraph/meta_handler/images.py
@@ -6,7 +6,7 @@ from psycopg2.sql import SQL, Identifier
 
 from splitgraph.connection import get_connection
 from splitgraph.constants import SPLITGRAPH_META_SCHEMA
-from splitgraph.constants import SplitGraphException
+from splitgraph.exceptions import SplitGraphException
 from splitgraph.meta_handler.common import select, insert
 from splitgraph.meta_handler.objects import get_full_object_tree
 from splitgraph.meta_handler.tables import get_object_for_table
