+++ test/splitgraph/conftest.py
@@ -4,9 +4,9 @@ import pytest
 
 from splitgraph.commands import *
 from splitgraph.commands import unmount, commit
-from splitgraph.commands.misc import make_conn, cleanup_objects
-from splitgraph.connection import get_connection, override_driver_connection
-from splitgraph.constants import PG_USER, PG_PWD, PG_DB, serialize_connection_string, to_repository as R, Repository
+from splitgraph.commands.misc import cleanup_objects
+from splitgraph.connection import get_connection, override_driver_connection, serialize_connection_string, make_conn
+from splitgraph.constants import PG_USER, PG_PWD, PG_DB, to_repository as R, Repository
 from splitgraph.meta_handler.common import setup_registry_mode, ensure_metadata_schema, toggle_registry_rls
 from splitgraph.meta_handler.misc import get_current_repositories
 from splitgraph.meta_handler.tags import set_tag, get_current_head
